Revert "Add component processor"
This reverts commit 78e4dcf2f032f68ef96e56c7714557d7919b361b.
