+++ packages/core/src/html/NodeProcessor.js
@@ -14,7 +14,6 @@ const { renderSiteNav } = require('./siteNavProcessor');
 const { processInclude, processPanelSrc } = require('./includePanelProcessor');
 const { Context } = require('./Context');
 const linkProcessor = require('./linkProcessor');
-const { ComponentProcessor } = require('./ComponentProcessor');
 const { insertTemporaryStyles } = require('./tempStyleProcessor');
 const { highlightCodeBlock } = require('./codeblockProcessor');
 
@@ -51,16 +50,15 @@ class NodeProcessor {
 
     // Store footnotes of <include>s and the main content, then combine them at the end
     this.renderedFootnotes = [];
-    this.componentProcessor = new ComponentProcessor(NodeProcessor);
   }
 
-  static renderMd(text) {
+  _renderMd(text) {
     return md.render(text, this.docId
       ? { docId: `${this.baseDocId}${this.docId}` }
       : { docId: this.baseDocId });
   }
 
-  static renderMdInline(text) {
+  _renderMdInline(text) {
     return md.renderInline(text, this.docId
       ? { docId: `${this.baseDocId}${this.docId}` }
       : { docId: this.baseDocId });
@@ -135,27 +133,6 @@ class NodeProcessor {
     delete node.attribs[attribute];
   }
 
-  static processAttributeWithoutOverride(node, attribute, isInline, slotName = attribute) {
-    const hasAttributeSlot = node.children
-        && node.children.some(child => NodeProcessor.getVslotShorthandName(child) === slotName);
-
-    if (!hasAttributeSlot && _.has(node.attribs, attribute)) {
-      let rendered;
-      if (isInline) {
-        rendered = NodeProcessor.renderMdInline(node.attribs[attribute]);
-      } else {
-        rendered = NodeProcessor.renderMd(node.attribs[attribute]);
-      }
-
-      const attributeSlotElement = cheerio.parseHTML(
-        `<template #${slotName}>${rendered}</template>`, true);
-      node.children
-          = node.children ? attributeSlotElement.concat(node.children) : attributeSlotElement;
-    }
-
-    delete node.attribs[attribute];
-  }
-
   /**
    * Takes an element, looks for direct elements with slots and transforms to avoid Vue parsing.
    * This is so that we can use bootstrap-vue popovers, tooltips, and modals.
@@ -197,6 +174,15 @@ class NodeProcessor {
       ...currentFrontMatter,
     };
     cheerio(node).remove();
+  }
+
+  /*
+   * Panels
+   */
+
+  _processPanelAttributes(node) {
+    this._processAttributeWithoutOverride(node, 'alt', false, '_alt');
+    this._processAttributeWithoutOverride(node, 'header', false);
   }
 
   /**
@@ -287,6 +273,48 @@ class NodeProcessor {
       });
   }
 
+  /*
+   * Questions, QOption, and Quizzes
+   */
+
+  _processQuestion(node) {
+    this._processAttributeWithoutOverride(node, 'header', false, 'header');
+    this._processAttributeWithoutOverride(node, 'hint', false, 'hint');
+    this._processAttributeWithoutOverride(node, 'answer', false, 'answer');
+  }
+
+  _processQOption(node) {
+    this._processAttributeWithoutOverride(node, 'reason', false, 'reason');
+  }
+
+  _processQuiz(node) {
+    this._processAttributeWithoutOverride(node, 'intro', false, 'intro');
+  }
+
+  /*
+   * Popovers, tooltips, modals, triggers
+   *
+   * We use bootstrap-vue's popovers, tooltips and modals, but perform various transformations
+   * to conform with our syntax instead, and to support triggers.
+   *
+   * For tooltips and popovers,
+   * The content / title is put into hidden [data-mb-slot-name] slots.
+   * Then, we call the relevant content getters inside core-web/index.js at runtime to get this content.
+   *
+   * For modals,
+   * only syntactic transformations are performed.
+   *
+   * For triggers,
+   * When building the site, we can't immediately tell whether a trigger references
+   * a modal, popover, or tooltip, as the element may not have been processed yet.
+   *
+   * So, we make every trigger try all 3 at runtime in the browser. (refer to Trigger.vue)
+   * It will attempt to open a modal, then a tooltip or popover.
+   * For modals, we simply attempt to show the modal via bootstrap-vue's programmatic api.
+   * The content of tooltips and popovers is retrieved from the [data-mb-slot-name] slots,
+   * then the <b-popover/tooltip> component is dynamically created appropriately.
+   */
+
   static addTriggerClass(node, trigger) {
     const triggerClass = trigger === 'click' ? 'trigger-click' : 'trigger';
     node.attribs.class = node.attribs.class ? `${node.attribs.class} ${triggerClass}` : triggerClass;
@@ -333,6 +361,18 @@ class NodeProcessor {
     });
   }
 
+  _processTooltip(node) {
+    this._processAttributeWithoutOverride(node, 'content', true, '_content');
+
+    node.name = 'span';
+    const trigger = node.attribs.trigger || 'hover';
+    const placement = node.attribs.placement || 'top';
+    node.attribs['data-mb-component-type'] = 'tooltip';
+    node.attribs[`v-b-tooltip.${trigger}.${placement}.html`] = 'tooltipInnerContentGetter';
+    NodeProcessor.addTriggerClass(node, trigger);
+    NodeProcessor._transformSlottedComponents(node);
+  }
+
   static _renameSlot(node, originalName, newName) {
     if (node.children) {
       node.children.forEach((child) => {
@@ -351,6 +391,139 @@ class NodeProcessor {
       node.attribs[newAttribute] = node.attribs[originalAttribute];
       delete node.attribs[originalAttribute];
     }
+  }
+
+  _processModalAttributes(node) {
+    NodeProcessor._warnDeprecatedAttributes(node, { title: 'header' });
+    NodeProcessor._warnDeprecatedSlotNames(node, {
+      'modal-header': 'header',
+      'modal-footer': 'footer',
+    });
+
+    this._processAttributeWithoutOverride(node, 'header', true, 'modal-title');
+    this._processAttributeWithoutOverride(node, 'title', true, 'modal-title');
+
+    NodeProcessor._renameSlot(node, 'header', 'modal-header');
+    NodeProcessor._renameSlot(node, 'footer', 'modal-footer');
+
+    node.name = 'b-modal';
+
+    NodeProcessor._renameAttribute(node, 'ok-text', 'ok-title');
+    NodeProcessor._renameAttribute(node, 'center', 'centered');
+
+    const hasOkTitle = _.has(node.attribs, 'ok-title');
+    const hasFooter = node.children
+      .some(child => NodeProcessor.getVslotShorthandName(child) === 'modal-footer');
+
+    if (!hasFooter && !hasOkTitle) {
+      // markbind doesn't show the footer by default
+      node.attribs['hide-footer'] = '';
+    } else if (hasOkTitle) {
+      // bootstrap-vue default is to show ok and cancel
+      // if there's an ok-title, markbind only shows the OK button.
+      node.attribs['ok-only'] = '';
+    }
+
+    if (node.attribs.backdrop === 'false') {
+      node.attribs['no-close-on-backdrop'] = '';
+    }
+    delete node.attribs.backdrop;
+
+    let size = '';
+    if (_.has(node.attribs, 'large')) {
+      size = 'lg';
+      delete node.attribs.large;
+    } else if (_.has(node.attribs, 'small')) {
+      size = 'sm';
+      delete node.attribs.small;
+    }
+    node.attribs.size = size;
+
+    // default for markbind is zoom, default for bootstrap-vue is fade
+    const effect = node.attribs.effect === 'fade' ? '' : 'mb-zoom';
+    node.attribs['modal-class'] = effect;
+
+    if (_.has(node.attribs, 'id')) {
+      node.attribs.ref = node.attribs.id;
+    }
+  }
+
+  /*
+   * Tabs
+   */
+
+  _processTabAttributes(node) {
+    this._processAttributeWithoutOverride(node, 'header', true, 'header');
+  }
+
+  /*
+   * Tip boxes
+   */
+
+  _processBoxAttributes(node) {
+    NodeProcessor._warnConflictingAttributes(node, 'light', ['seamless']);
+    NodeProcessor._warnConflictingAttributes(node, 'no-background', ['background-color', 'seamless']);
+    NodeProcessor._warnConflictingAttributes(node, 'no-border',
+                                             ['border-color', 'border-left-color', 'seamless']);
+    NodeProcessor._warnConflictingAttributes(node, 'no-icon', ['icon']);
+    NodeProcessor._warnDeprecatedAttributes(node, { heading: 'header' });
+
+    this._processAttributeWithoutOverride(node, 'icon', true, 'icon');
+    this._processAttributeWithoutOverride(node, 'header', false, 'header');
+
+    this._processAttributeWithoutOverride(node, 'heading', false, 'header');
+  }
+
+  /*
+   * Dropdowns
+   */
+
+  _processDropdownAttributes(node) {
+    const hasHeaderSlot = node.children
+      && node.children.some(child => NodeProcessor.getVslotShorthandName(child) === 'header');
+
+    // If header slot is present, the header attribute has no effect, and we can simply remove it.
+    if (hasHeaderSlot) {
+      if (_.has(node.attribs, 'header')) {
+        logger.warn(`${node.name} has a header slot, 'header' attribute has no effect.`);
+      }
+      if (_.has(node.attribs, 'text')) {
+        logger.warn(`${node.name} has a header slot, 'text' attribute has no effect.`);
+      }
+      delete node.attribs.header;
+      delete node.attribs.text;
+      return;
+    }
+
+    NodeProcessor._warnDeprecatedAttributes(node, { text: 'header' });
+    NodeProcessor._warnConflictingAttributes(node, 'header', ['text']);
+    // header attribute takes priority over text attribute if both 'text' and 'header' is used
+    if (_.has(node.attribs, 'header')) {
+      this._processAttributeWithoutOverride(node, 'header', true, 'header');
+      delete node.attribs.text;
+    } else {
+      this._processAttributeWithoutOverride(node, 'text', true, 'header');
+    }
+  }
+
+  /**
+   * Thumbnails
+   */
+
+  _processThumbnailAttributes(node) {
+    const isImage = _.has(node.attribs, 'src') && node.attribs.src !== '';
+    if (isImage) {
+      return;
+    }
+
+    const text = _.has(node.attribs, 'text') ? node.attribs.text : '';
+    if (text === '') {
+      return;
+    }
+
+    const renderedText = this._renderMdInline(text);
+    node.children = cheerio.parseHTML(renderedText);
+    delete node.attribs.text;
   }
 
   /**
@@ -515,43 +688,41 @@ class NodeProcessor {
       case 'include':
         this.docId += 1; // used in markdown-it-footnotes
         return processInclude(node, context, this.pageSources, this.variableProcessor,
-                              text => NodeProcessor.renderMd(text),
-                              text => NodeProcessor.renderMdInline(text),
+                              text => this._renderMd(text), text => this._renderMdInline(text),
                               this.config);
       case 'panel':
-        this.componentProcessor.processPanelAttributes(node);
+        this._processPanelAttributes(node);
         return processPanelSrc(node, context, this.pageSources, this.config);
       case 'question':
-        this.componentProcessor._processQuestion(node);
+        this._processQuestion(node);
         break;
       case 'q-option':
-        this.componentProcessor._processQOption(node);
+        this._processQOption(node);
         break;
       case 'quiz':
-        this.componentProcessor._processQuiz(node);
+        this._processQuiz(node);
         break;
       case 'popover':
-        // this.componentProcessor._processPopover(node);
         this._processPopover(node);
         break;
       case 'tooltip':
-        this.componentProcessor._processTooltip(node);
+        this._processTooltip(node);
         break;
       case 'modal':
-        this.componentProcessor._processModalAttributes(node);
+        this._processModalAttributes(node);
         break;
       case 'tab':
       case 'tab-group':
-        this.componentProcessor._processTabAttributes(node);
+        this._processTabAttributes(node);
         break;
       case 'box':
-        this.componentProcessor._processBoxAttributes(node);
+        this._processBoxAttributes(node);
         break;
       case 'dropdown':
-        this.componentProcessor._processDropdownAttributes(node);
+        this._processDropdownAttributes(node);
         break;
       case 'thumbnail':
-        this.componentProcessor._processThumbnailAttributes(node);
+        this._processThumbnailAttributes(node);
         break;
       case 'annotation':
         NodeProcessor._processAnnotationAttributes(node);
@@ -630,11 +801,11 @@ class NodeProcessor {
     switch (node.name) {
     case 'md':
       node.name = 'span';
-      node.children = cheerio.parseHTML(NodeProcessor.renderMdInline(cheerio.html(node.children)), true);
+      node.children = cheerio.parseHTML(this._renderMdInline(cheerio.html(node.children)), true);
       break;
     case 'markdown':
       node.name = 'div';
-      node.children = cheerio.parseHTML(NodeProcessor.renderMd(cheerio.html(node.children)), true);
+      node.children = cheerio.parseHTML(this._renderMd(cheerio.html(node.children)), true);
       break;
     default:
       break;
@@ -693,7 +864,7 @@ class NodeProcessor {
       const parser = new htmlparser.Parser(handler);
       const fileExt = utils.getExt(file);
       if (utils.isMarkdownFileExt(fileExt)) {
-        const renderedContent = NodeProcessor.renderMd(content);
+        const renderedContent = this._renderMd(content);
         // Wrap with <root> as $.remove() does not work on top level nodes
         parser.parseComplete(`<root>${renderedContent}</root>`);
       } else if (fileExt === 'html') {
